+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/config
 
+## 1.2.5
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+
 ## 1.2.4
 
 ### Patch Changes