+++ packages/context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/context
 
+## 1.11.0
+
+### Minor Changes
+
+-   Add support for Application profiles
+
 ## 1.10.0
 
 ### Minor Changes