+++ packages/di/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/di
 
+## 1.3.0
+
+### Minor Changes
+
+-   Add support for Application profiles
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+
 ## 1.2.4
 
 ### Patch Changes