+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/starter-firebase
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+    -   @nodeboot/engine@1.4.0
+    -   @nodeboot/core@1.13.0
+    -   @nodeboot/config@1.2.5
+
 ## 1.1.9
 
 ### Patch Changes