+++ starters/persistence/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/starter-persistence
 
+## 1.9.0
+
+### Minor Changes
+
+-   Add support for Application profiles
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+    -   @nodeboot/core@1.13.0
+    -   @nodeboot/di@1.3.0
+    -   @nodeboot/config@1.2.5
+
 ## 1.8.10
 
 ### Patch Changes