+++ starters/scheduler/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/starter-scheduler
 
+## 1.4.0
+
+### Minor Changes
+
+-   Add support for Application profiles
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+
 ## 1.3.4
 
 ### Patch Changes