+++ tests/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/test
 
+## 1.2.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.11.0
+    -   @nodeboot/core@1.13.0
+    -   @nodeboot/config@1.2.5
+
 ## 1.2.9
 
 ### Patch Changes